Gestions albums et début interface élève
Suite de la gestions des albums photos (modification/suppression des
albums).
Travail sur l'interface élève des projets.
