+++ src/CEC/SecteurProjetsBundle/Entity/Projet.php~
@@ -89,7 +89,7 @@ class Projet
 	*
 	* @ORM\Column(name="inscriptions_ouvertes", type="boolean")
 	*/
-	private $inscriptionsOuvertes;
+	private $inscriptionsOuvertes = false;
 	
 	/**
 	* Liste des réunions liées au projet
@@ -110,9 +110,23 @@ class Projet
 	/**
 	* @var \Doctrine\Common\Collections\Collection
 	*
-	* @ORM\ManyToMany(targetEntity="\CEC\MembreBundle\Entity\Membre", mappedBy="contactProjets")
+	* @ORM\ManyToMany(targetEntity="\CEC\MembreBundle\Entity\Membre", inversedBy="contactProjets")
 	*/
-	private $contacts;
+	private $contacts =array();
+	
+	/**
+	* @var \Doctrine\Common\Collections\Collection
+	*
+	*@ORM\OneToMany(targetEntity="\CEC\SecteurProjetsBundle\Entity\Album", mappedBy="projet")
+	*/
+	private $albums;
+
+    /**
+    * @var \Doctrine\Common\Collections\Collection
+    *
+    *@ORM\ManyToMany(targetEntity="\CEC\MembreBundle\Entity\Eleve", mappedBy="projets")
+    */
+    private $inscrits;
 
 
     /**
@@ -317,7 +331,19 @@ class Projet
      */
     public function setInscriptionsOuvertes($etat)
     {
-        $this->inscriptionsOuvertes = $inscriptionsOuvertes;
+        $this->inscriptionsOuvertes = $etat;
+    
+        return $this;
+    }
+	
+	/**
+     * Set inscriptionsOuvertes
+     *
+     * @return Projet
+     */
+    public function switchInscriptionsOuvertes()
+    {
+        $this->inscriptionsOuvertes = !$this->inscriptionsOuvertes;
     
         return $this;
     }
@@ -437,5 +463,102 @@ class Projet
     public function getLieu()
     {
         return $this->lieu;
+    }
+    /**
+     * Constructor
+     */
+    public function __construct()
+    {
+        $this->reunions = new \Doctrine\Common\Collections\ArrayCollection();
+        $this->contacts = new \Doctrine\Common\Collections\ArrayCollection();
+    }
+    
+    /**
+     * Add reunions
+     *
+     * @param \CEC\SecteurProjetsBundle\Entity\Reunion $reunions
+     * @return Projet
+     */
+    public function addReunion(\CEC\SecteurProjetsBundle\Entity\Reunion $reunions)
+    {
+        $this->reunions[] = $reunions;
+    
+        return $this;
+    }
+
+    /**
+     * Remove reunions
+     *
+     * @param \CEC\SecteurProjetsBundle\Entity\Reunion $reunions
+     */
+    public function removeReunion(\CEC\SecteurProjetsBundle\Entity\Reunion $reunions)
+    {
+        $this->reunions->removeElement($reunions);
+    }
+	
+	/**
+     * Add album
+     *
+     * @param \CEC\SecteurProjetsBundle\Entity\Album $album
+     * @return Membre
+     */
+    public function addAlbum(\CEC\SecteurProjetsBundle\Entity\Album $album)
+    {
+        $this->albums[] = $album;
+
+        return $this;
+    }
+
+    /**
+     * Remove album
+     *
+     * @param \CEC\SecteurProjetsBundle\Entity\Album $album
+     */
+    public function removeAlbum(\CEC\SecteurProjetsBundle\Entity\Album $album)
+    {
+        $this->albums->removeElement($album);
+    }
+
+    /**
+     * Get albums
+     *
+     * @return \Doctrine\Common\Collections\Collection
+     */
+    public function getAlbums()
+    {
+        return $this->albums;
+    }
+
+    /**
+     * Add inscrits
+     *
+     * @param \CEC\MembreBundle\Entity\Eleve $inscrits
+     * @return Projet
+     */
+    public function addInscrit(\CEC\MembreBundle\Entity\Eleve $inscrits)
+    {
+        $this->inscrits[] = $inscrits;
+    
+        return $this;
+    }
+
+    /**
+     * Remove inscrits
+     *
+     * @param \CEC\MembreBundle\Entity\Eleve $inscrits
+     */
+    public function removeInscrit(\CEC\MembreBundle\Entity\Eleve $inscrits)
+    {
+        $this->inscrits->removeElement($inscrits);
+    }
+
+    /**
+     * Get inscrits
+     *
+     * @return \Doctrine\Common\Collections\Collection 
+     */
+    public function getInscrits()
+    {
+        return $this->inscrits;
     }
 }
