+++ src/CEC/SecteurProjetsBundle/Resources/config/routing/modification.yml
@@ -1,23 +1,40 @@
+
+# Routes d'édition des projets
 editer_projet:
-    pattern: /editer/{slug}
+    pattern: /editer-projet/{slug}
     defaults: { _controller: CECSecteurProjetsBundle:Projets:editer}
-    
+
+
+#Routes d'édition des réunions d'information
 creer_reunion:
     pattern: /creer-reunion
-    defaults: { _controller: CECSecteurProjetsBundle:Projets:creerReunion}
+    defaults: { _controller: CECSecteurProjetsBundle:Reunions:creerReunion}
 
 modifier_reunion:
     pattern: /editer-reunion/{id}
-    defaults: {_controller: CECSecteurProjetsBundle:Projets:modifierReunion}
+    defaults: {_controller: CECSecteurProjetsBundle:Reunions:modifierReunion}
     requirements:
         id: \d+
 
 supprimer_reunion:
     pattern: /supprimer-reunion/{id}
-    defaults: {_controller:CECSecteurProjetsBundle:Projets:supprimerReunion}
+    defaults: {_controller:CECSecteurProjetsBundle:Reunions:supprimerReunion}
     requirements:
         id: \d+
 
+inscription_reunion:
+    pattern: /inscription-reunion/{id}
+    defaults: {_controller: CECSecteurProjetsBundle:Reunions:inscriptionReunion}
+    requirements:
+        id: \d+
+
+desinscription_reunion:
+    pattern: /desinscription-reunion/{id}
+    defaults: {_controller: CECSecteurProjetsBundle:Reunions:desinscriptionReunion}
+    requirements:
+        id: \d+
+
+# Routes de gestion des inscriptions
 dossier_inscription:
     pattern: /dossier-inscription
     defaults: {_controller: CECSecteurProjetsBundle:Projets:uploadDossier}
@@ -25,7 +42,20 @@ dossier_inscription:
 ouverture_inscription:
     pattern: /ouverture-inscriptions
     defaults: {_controller: CECSecteurProjetsBundle:Projets:inscriptions}
-    
+
+
+# Routes d'édition des albums photos
 ajouter_photos:
-    pattern: /ajouter-photos
-    defaults: {_controller: CECSecteurProjetsBundle:Projets:ajouterPhotos}
+    pattern: /ajouter-photos/{id}
+    defaults: {_controller: CECSecteurProjetsBundle:Albums:ajouterPhotos, id: 0}
+
+gerer_albums:
+    pattern: /editer-albums
+    defaults: {_controller: CECSecteurProjetsBundle:Albums:gererAlbums}
+
+gerer_album:
+    pattern: /editer-albums/{id}
+    defaults: {_controller: CECSecteurProjetsBundle:Albums:gererAlbum, image_id: 0}
+    requirements:
+        id: \d+
+        image_id: \d+
