+++ src/CEC/SecteurProjetsBundle/Resources/views/Albums/gererAlbums.html.twig
@@ -1,0 +1,39 @@
+{% extends 'CECSecteurProjetsBundle::base.html.twig' %}
+
+
+
+{% block right %}
+
+{{parent()}}
+
+<h1> Gérer les albums Photos </h1>
+<br/>
+
+<h3> Créer un album </h3>
+
+<a href="{{path('ajouter_photos')}}" class="btn btn-default"> Créer un album </a><br/>
+
+<h3> Quel album voulez-vous modifier ? </h3>
+
+<table class="table table-bordered table-striped table-large">
+    <tbody>
+	{% for projet in projets%}
+        <tr>
+            <th>
+                {{projet.nom}}
+            </th>
+			{% for album in projet.albums %}
+				<td>
+					<a href="{{path('gerer_album', {'id': album.id})}}">{{album.annee}}</a>
+				</td>
+			{% else %}
+				<td> 
+					<em>Pas encore d'albums disponibles</em>
+				</td>
+			{% endfor %}
+        </tr>
+	{% endfor %}
+    </tbody>
+</table>
+
+{% endblock %}
