+++ src/CEC/SecteurProjetsBundle/Resources/views/Albums/voirPhotos.html.twig
@@ -1,0 +1,45 @@
+{% extends 'CECSecteurProjetsBundle::base.html.twig' %}
+
+
+
+{% block right %}
+
+{{parent()}}
+<div class="well" style = "padding-left:20px;padding-right:15px;">
+	<h1>{{album.projet.nom}} - Edition {{album.annee}}</h1>
+
+	<div id="carousel-example-generic" class="carousel slide" data-ride="carousel">
+	  <!-- Indicators -->
+	  <ol class="carousel-indicators">
+		<li data-target="#carousel-example-generic" data-slide-to="0" class="active"></li>
+		{% for i in range(1, (album.images|length)-1) %}
+			<li data-target="#carousel-example-generic" data-slide-to="{{i}}"></li>
+		{% endfor %}
+	  </ol>
+
+	  <!-- Wrapper for slides -->
+	  <div class="carousel-inner" role="listbox">
+		{% for image in album.images%}
+			<div class="item {% if loop.first %}active {% endif %}">
+			  <img src="{{asset(image.webPath)}}" alt="{{image.alt}}" style="margin:auto;">
+			  <div class="carousel-caption">
+				{{image.legende}}
+			  </div>
+			</div>
+		{% endfor %}
+	  </div>
+
+	  <!-- Controls -->
+	  <a class="left carousel-control" href="#carousel-example-generic" role="button" data-slide="prev">
+		<span class="glyphicon glyphicon-chevron-left" aria-hidden="true"></span>
+		<span class="sr-only"><</span>
+	  </a>
+	  <a class="right carousel-control" href="#carousel-example-generic" role="button" data-slide="next">
+		<span class="glyphicon glyphicon-chevron-right" aria-hidden="true"></span>
+		<span class="sr-only">></span>
+	  </a>
+	</div>
+	
+	<a href="{{path('voir_albums')}}" class="btn btn-default">Retourner à la liste des albums</a>
+</div>
+{% endblock %}
