+++ src/CEC/SecteurProjetsBundle/Resources/views/Projets/menu.html.twig
@@ -58,11 +58,13 @@
 			<li> <a href="{{path('dossier_inscription')}}">Dossiers d'inscription</a></li>
 			<li> <a href="{{path('ouverture_inscription')}}">Ouvrir les inscriptions</a></li>
 			<li> <a href="{{path('voir_albums')}}">Photos des voyages</a></li>
-			<li> <a href="{{path('ajouter_photos')}}"> Ajouter des photos</a></li>
+			<li> <a href="{{path('gerer_albums')}}"> Gérer les albums</a></li>
 		</ul>
 	{% elseif is_granted('ROLE_ELEVE')%}
 		<ul class="nav nav-pills nav-stacked">
-			<li> Menu d'inscription </li>
+			<li> <a href="{{path('suivi_inscriptions')}}">Suivi de tes inscriptions </a> </li>
+			<li> <a href="{{path('liste_reunions')}}"> Réunions d'information </a> </li>
+			<li> <a href="{{path('voir_albums')}}">Photos des voyages</a></li>
 		</ul>
 	{% elseif is_granted('ROLE_PROFESSEUR_REFERENT') %}
 		<ul class="nav nav-pills nav-stacked">
