+++ src/CEC/SecteurProjetsBundle/Resources/views/Projets/uploadDossier.html.twig
@@ -1,4 +1,4 @@
-{% extends 'CECSecteurProjetsBundle:Projets:base.html.twig' %}
+{% extends 'CECSecteurProjetsBundle::base.html.twig' %}
 
 {% block right %}
