+++ src/CEC/SecteurProjetsBundle/Resources/views/Reunions/modifierReunion.html.twig
@@ -1,5 +1,5 @@
 
-{% extends 'CECSecteurProjetsBundle:Projets:base.html.twig' %}
+{% extends 'CECSecteurProjetsBundle::base.html.twig' %}
 
 {% block right %}
 {{parent()}}
