+++ src/CEC/SecteurProjetsBundle/Resources/views/Reunions/voirReunion.html.twig
@@ -1,4 +1,4 @@
-{% extends 'CECSecteurProjetsBundle:Projets:base.html.twig' %}
+{% extends 'CECSecteurProjetsBundle::base.html.twig' %}
 
 {% block right %}
 {{parent()}}
@@ -34,12 +34,24 @@
 			<td>
 				{{reunion.description}}
 			</td>
-			<td>
-				<a href="{{path('modifier_reunion', {'id':reunion.id})}}" class="btn btn-default">Modifier</a>
-			</td>
-			<td>
-				<a href="{{path('supprimer_reunion', {'id':reunion.id})}}" class="btn btn-danger">Supprimer</a>
-			</td>
+			{% if is_granted('ROLE_SECTEUR_PROJETS') %}
+				<td>
+					{{reunion.presents|length}} inscrits
+				<td>
+					<a href="{{path('modifier_reunion', {'id':reunion.id})}}" class="btn btn-default">Modifier</a>
+				</td>
+				<td>
+					<a href="{{path('supprimer_reunion', {'id':reunion.id})}}" class="btn btn-danger">Supprimer</a>
+				</td>
+			{% elseif is_granted('ROLE_ELEVE') %}
+				<td>
+					{% if app.user in reunion.presents %}
+						<em> Tu es déjà inscrit à cette réunion !</em>
+					{% else %}
+						<a href="{{path('inscription_reunion', {'id': reunion.id})}}" class="btn btn-default">S'inscrire</a>
+					{% endif %}
+				</td>
+			{% endif %}
 		</tr>		
 		{% endif %}
 		{% endfor %} 
@@ -49,7 +61,7 @@
 
 <h3>Créer une nouvelle réunion d'information</h3>
 
-{% render('CECSecteurProjetsBundle:Projets:creerReunion')%}
+{% render('CECSecteurProjetsBundle:Reunions:creerReunion')%}
 
 </div>
