+++ src/CEC/SecteurProjetsBundle/Resources/views/base.html.twig
@@ -1,0 +1,14 @@
+{% extends 'CECMainBundle::2-columns.html.twig' %}
+
+{% block stylesheets %}
+    {{ parent() }}
+    {% stylesheets
+        '@CECTutoratBundle/Resources/public/less/cordees.less'
+    %}
+    <link href="{{ asset_url }}" rel="stylesheet" media="screen" />
+    {% endstylesheets %}
+{% endblock %}
+
+{% block left %}
+    {% render 'CECSecteurProjetsBundle:Projets:menu' with { 'request': app.request } %}
+{% endblock %}
